Move AccessibilityRenderingTest cases into sample module
