+++ paparazzi-gradle-plugin/src/test/projects/accessibility-rendering/src/test/java/app/cash/paparazzi/plugin/test/AccessibilityRenderingTest.kt
@@ -1,62 +1,41 @@
 package app.cash.paparazzi.plugin.test
 
-import android.content.Context
-import android.text.Editable
-import android.view.View
-import android.view.View.GONE
-import android.widget.Button
-import android.widget.EditText
+import android.view.ViewGroup.LayoutParams.MATCH_PARENT
+import android.view.ViewGroup.LayoutParams.WRAP_CONTENT
 import android.widget.LinearLayout
-import android.widget.ListView
-import android.widget.TextView
-import androidx.compose.foundation.background
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.wrapContentSize
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.progressSemantics
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.DropdownMenu
-import androidx.compose.material3.DropdownMenuItem
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.ModalBottomSheet
-import androidx.compose.material3.SheetState
-import androidx.compose.material3.SheetValue
-import androidx.compose.material3.Slider
-import androidx.compose.material3.Text
-import androidx.compose.ui.ExperimentalComposeUiApi
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.grid.GridCells
+import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
+import androidx.compose.foundation.selection.toggleable
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Add
+import androidx.compose.material3.Button
+import androidx.compose.material3.Checkbox
+import androidx.compose.material3.IconToggleButton
+import androidx.compose.material3.TextField
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.alpha
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.ComposeView
-import androidx.compose.ui.platform.LocalDensity
-import androidx.compose.ui.semantics.CustomAccessibilityAction
-import androidx.compose.ui.semantics.LiveRegionMode
-import androidx.compose.ui.semantics.clearAndSetSemantics
+import androidx.compose.ui.semantics.Role
 import androidx.compose.ui.semantics.contentDescription
-import androidx.compose.ui.semantics.customActions
-import androidx.compose.ui.semantics.invisibleToUser
-import androidx.compose.ui.semantics.liveRegion
+import androidx.compose.ui.semantics.heading
+import androidx.compose.ui.semantics.role
+import androidx.compose.ui.semantics.selected
 import androidx.compose.ui.semantics.semantics
-import androidx.compose.ui.semantics.setProgress
-import androidx.compose.ui.semantics.traversalIndex
-import androidx.compose.ui.text.LinkAnnotation
-import androidx.compose.ui.text.buildAnnotatedString
+import androidx.compose.ui.semantics.stateDescription
 import androidx.compose.ui.unit.dp
-import androidx.core.view.AccessibilityDelegateCompat
-import androidx.core.view.ViewCompat
-import androidx.core.view.accessibility.AccessibilityNodeInfoCompat
 import app.cash.paparazzi.DeviceConfig
 import app.cash.paparazzi.Paparazzi
 import app.cash.paparazzi.accessibility.AccessibilityRenderExtension
-import com.android.ide.common.rendering.api.SessionParams.RenderingMode
 import org.junit.Rule
 import org.junit.Test
 
-@OptIn(ExperimentalComposeUiApi::class)
 class AccessibilityRenderingTest {
   @get:Rule
   val paparazzi = Paparazzi(
@@ -66,274 +45,32 @@ class AccessibilityRenderingTest {
   )
 
   @Test
-  fun `mixed compose usage`() {
-    val mixedView = MixedView(paparazzi.context)
-    paparazzi.snapshot(mixedView)
-  }
+  fun multiComposeViews() {
+    val view = LinearLayout(paparazzi.context).apply {
+      orientation = LinearLayout.VERTICAL
+      layoutParams = LinearLayout.LayoutParams(MATCH_PARENT, MATCH_PARENT)
+      setPaddingRelative(25, 25, 25, 25)
 
-  @Test
-  fun dropDownMaterial3() {
-    paparazzi.snapshot {
-      Box(Modifier.fillMaxSize()) {
-        DropdownMenu(
-          expanded = true,
-          onDismissRequest = { }
-        ) {
-          DropdownMenuItem(
-            text = {
-              Text(
-                text = "Label 1"
-              )
-            },
-            onClick = {}
-          )
-          DropdownMenuItem(
-            text = {
-              Text(
-                text = "Label 2"
-              )
-            },
-            onClick = {}
-          )
-          DropdownMenuItem(
-            text = {
-              Text(
-                text = "Label 3"
-              )
-            },
-            onClick = {}
-          )
-        }
-      }
-    }
-  }
-
-  @Test
-  @OptIn(ExperimentalMaterial3Api::class)
-  fun modalBottomSheetMaterial3() {
-    paparazzi.snapshot {
-      ModalBottomSheet(
-        onDismissRequest = {},
-        sheetState = SheetState(
-          skipPartiallyExpanded = true,
-          density = LocalDensity.current,
-          initialValue = SheetValue.Expanded
-        )
-      ) {
-        Text(text = "Text 2")
-      }
-      Text(modifier = Modifier.wrapContentSize(), text = "Text 1")
-    }
-  }
-
-  @Test
-  fun `verify changing view hierarchy order doesn't change accessibility colors`() {
-    val mixedView = MixedView(paparazzi.context).apply {
       addView(
-        ComposeView(context).apply {
-          id = 10
-          setContent {
-            Box(modifier = Modifier.size(50.dp)) {}
-          }
+        ComposeView(context = paparazzi.context).apply {
+          setPaddingRelative(32, 32, 32, 32)
+          setContent { androidx.compose.material.Text("Number 1") }
         },
-        0,
-        LinearLayout.LayoutParams(0, 0)
+        LinearLayout.LayoutParams(WRAP_CONTENT, 0, 1f)
       )
-    }
-    paparazzi.snapshot(mixedView)
-  }
-
-  @Test
-  fun `verify clear and set semantics`() {
-    paparazzi.snapshot {
-      Box(modifier = Modifier.clickable {}) {
-        Column(
-          modifier = Modifier.clearAndSetSemantics {
-            contentDescription = "OVERRIDDEN CONTENT DESCRIPTION"
-          }
-        ) {
-          Text(text = "Text")
-        }
-      }
-    }
-  }
-
-  @Test
-  fun `verify hidden ComposeView content is not in legend`() {
-    val view = ComposeView(paparazzi.context).apply {
-      visibility = GONE
-      setContent {
-        Column {
-          Text(text = "Text 1")
-          Text(text = "Text 2")
-        }
-      }
-    }
-
-    paparazzi.snapshot(view)
-  }
-
-  @Test
-  fun `verify hidden views are not in legend`() {
-    val view = ComposeView(paparazzi.context).apply {
-      setContent {
-        Column {
-          Text(
-            modifier = Modifier
-              .semantics {
-                invisibleToUser()
-              },
-            text = "Text invisible to user"
-          )
-          Text(
-            modifier = Modifier
-              .alpha(0f),
-            text = "Text with zero alpha"
-          )
-          Text(text = "Text that is visible!")
-        }
-      }
-    }
-
-    paparazzi.snapshot(view)
-  }
-
-  @Test
-  fun legendDoesNotScale() {
-    paparazzi.unsafeUpdateConfig(deviceConfig = DeviceConfig.PIXEL.copy(fontScale = 2.0f))
-    paparazzi.snapshot {
-      Column(Modifier.background(Color.LightGray)) {
-        androidx.compose.material.Text("Some text that will appear scaled in the UI, but not scaled in the legend")
-      }
-    }
-  }
-
-  @Test(expected = IllegalStateException::class)
-  fun renderingModeSHRINKThrowsException() {
-    paparazzi.unsafeUpdateConfig(renderingMode = RenderingMode.SHRINK)
-    paparazzi.snapshot {
-      Column(Modifier.background(Color.LightGray)) {
-        Text("SHRINK and AccessibilityRenderExtension are not supported together")
-      }
-    }
-  }
-
-  @Test
-  fun `verify progress semantic information represented`() {
-    val view = ComposeView(paparazzi.context).apply {
-      setContent {
-        Column {
-          Slider(
-            value = 0.5f,
-            onValueChange = { _ -> },
-            valueRange = 0f..1f
-          )
-          CircularProgressIndicator(
-            modifier = Modifier.progressSemantics()
-          )
-          Slider(
-            modifier = Modifier.semantics {
-              setProgress("Adjust volume") { _ -> true }
-            },
-            value = 0.26f,
-            onValueChange = { _ -> },
-            valueRange = 0f..1f
-          )
-        }
-      }
-    }
-
-    paparazzi.snapshot(view)
-  }
-
-  @Test
-  fun `verify link annotation and custom actions`() {
-    val view = ComposeView(paparazzi.context).apply {
-      setContent {
-        Column {
-          // Test link annotation
-          val annotatedString = buildAnnotatedString {
-            append("Visit ")
-            pushLink(LinkAnnotation.Url("https://www.example.com"))
-            append("Url ")
-            pop()
-            pushLink(LinkAnnotation.Clickable(tag = "CLICK") {})
-            append("Clickable")
-            pop()
-          }
-
-          Text(text = annotatedString)
-
-          // Test custom actions
-          Box(
-            modifier = Modifier
-              .size(100.dp)
-              .background(Color.LightGray)
-              .semantics(mergeDescendants = true) {
-                customActions = listOf(
-                  CustomAccessibilityAction("Action 1") { true },
-                  CustomAccessibilityAction("Action 2") { true }
-                )
-              }
-          ) {
-            Text("Box with custom actions")
-          }
-        }
-      }
-    }
-
-    paparazzi.snapshot(view)
-  }
-
-  @Test
-  fun `verify view custom actions`() {
-    val view = buildViewWithCustomActions(paparazzi.context)
-    paparazzi.snapshot(view, name = "custom-actions")
-  }
-
-  @Test
-  fun `verify compose live region`() {
-    val view = ComposeView(paparazzi.context).apply {
-      setContent {
-        Column {
-          Box(
-            modifier = Modifier
-              .size(100.dp)
-              .background(Color.LightGray)
-              .semantics(mergeDescendants = true) {
-                liveRegion = LiveRegionMode.Polite
-              }
-          ) {
-            Text("Box with live region")
-          }
-        }
-      }
-    }
-
-    paparazzi.snapshot(view)
-  }
-
-  @Test
-  fun `verify view live region`() {
-    val view = LinearLayout(paparazzi.context).apply {
       addView(
-        TextView(paparazzi.context).apply {
-          text = "Live Region Text"
-          accessibilityLiveRegion = View.ACCESSIBILITY_LIVE_REGION_ASSERTIVE
-        }
+        ComposeView(context = paparazzi.context).apply {
+          setPaddingRelative(32, 32, 32, 32)
+          setContent { androidx.compose.material.Text("Number 2") }
+        },
+        LinearLayout.LayoutParams(WRAP_CONTENT, 0, 1f)
       )
-    }
-
-    paparazzi.snapshot(view)
-  }
-
-  @Test
-  fun `verify view EditText`() {
-    val view = LinearLayout(paparazzi.context).apply {
       addView(
-        EditText(context).apply {
-          text = Editable.Factory.getInstance().newEditable("Text input box")
-        }
+        ComposeView(context = paparazzi.context).apply {
+          setPaddingRelative(32, 32, 32, 32)
+          setContent { androidx.compose.material.Text("Number 3") }
+        },
+        LinearLayout.LayoutParams(WRAP_CONTENT, 0, 1f)
       )
     }
 
@@ -341,82 +78,140 @@ class AccessibilityRenderingTest {
   }
 
   @Test
-  fun `verify compose list semantics`() {
-    paparazzi.snapshot {
-      LazyColumn {
-        items(5) {
-          Text(text = "Item = $it")
-        }
-      }
-    }
-  }
-
-  @Test
-  fun `verify view list semantics`() {
-    val listView = ListView(paparazzi.context)
-    listView.adapter = object : android.widget.ArrayAdapter<String>(
-      paparazzi.context,
-      android.R.layout.simple_list_item_1,
-      listOf("Item = 0", "Item = 1", "Item = 2", "Item = 3", "Item = 4")
-    ) {
-      override fun getView(position: Int, convertView: View?, parent: android.view.ViewGroup): View {
-        val view = super.getView(position, convertView, parent)
-        view.contentDescription = "Item = $position"
-        return view
-      }
-    }
-
-    paparazzi.snapshot(listView)
-  }
-
-  @Test
-  fun `verify traversalIndex order`() {
+  fun compositeItems() {
     paparazzi.snapshot {
       Column {
-        Text(
-          text = "Third",
-          modifier = Modifier
-            .semantics { traversalIndex = 2f }
+        Row(
+          Modifier
+            .toggleable(
+              value = true,
+              role = Role.Checkbox,
+              onValueChange = { }
+            )
+            .fillMaxWidth()
+        ) {
+          androidx.compose.material.Text("Option", Modifier.weight(1f))
+          Checkbox(checked = true, onCheckedChange = null)
+        }
+        Box(
+          Modifier
+            .align(Alignment.CenterHorizontally)
+            .clickable(onClickLabel = "On Click Label") { }
         )
-        Text(
-          text = "First",
-          modifier = Modifier
-            .semantics { traversalIndex = -1f }
-        )
-        Text(
-          // Displayed second as the default traversalIndex is 0
-          text = "Second"
-        )
+        Row(modifier = Modifier.semantics(mergeDescendants = true) {}) {
+          Image(
+            imageVector = Icons.Filled.Add,
+            contentDescription = null // decorative
+          )
+          Column(modifier = Modifier.semantics(mergeDescendants = true) {}) {
+            androidx.compose.material.Text("Text")
+            androidx.compose.material.Text(
+              text = "more text",
+              modifier = Modifier.semantics { contentDescription = "custom description" }
+            )
+            Column(modifier = Modifier.semantics(mergeDescendants = true) {}) {
+              androidx.compose.material.Text("Nested text")
+              androidx.compose.material.Text(
+                text = "more text",
+                modifier = Modifier.semantics { contentDescription = "custom description" }
+              )
+            }
+          }
+        }
+        TextField(value = "Some text", label = {
+          androidx.compose.material.Text(text = "text field label")
+        }, onValueChange = {})
       }
     }
   }
 
-  private fun buildViewWithCustomActions(context: Context) =
-    LinearLayout(context).apply {
-      addView(
-        Button(context).apply {
-          text = "Actions Button"
-          ViewCompat.setAccessibilityDelegate(
-            this,
-            object : AccessibilityDelegateCompat() {
-              override fun onInitializeAccessibilityNodeInfo(host: View, info: AccessibilityNodeInfoCompat) {
-                super.onInitializeAccessibilityNodeInfo(host, info)
-                info.addAction(
-                  AccessibilityNodeInfoCompat.AccessibilityActionCompat(
-                    AccessibilityNodeInfoCompat.ACTION_CLICK,
-                    "Custom Click Action"
-                  )
-                )
-                info.addAction(
-                  AccessibilityNodeInfoCompat.AccessibilityActionCompat(
-                    AccessibilityNodeInfoCompat.ACTION_LONG_CLICK,
-                    "Custom Long Press Action"
-                  )
-                )
-              }
-            }
-          )
+  @Test
+  fun buttonStates() {
+    paparazzi.snapshot {
+      LazyVerticalGrid(columns = GridCells.Fixed(2)) {
+        item {
+          Button(
+            modifier = Modifier.padding(24.dp),
+            onClick = {}
+          ) {
+            androidx.compose.material.Text("01")
+          }
         }
-      )
+        item {
+          Button(
+            modifier = Modifier
+              .padding(24.dp),
+            onClick = {},
+            enabled = false
+          ) {
+            androidx.compose.material.Text("02")
+          }
+        }
+        item {
+          Button(
+            modifier = Modifier
+              .padding(24.dp)
+              .clickable(onClickLabel = "Explore") {},
+            onClick = {}
+          ) {
+            androidx.compose.material.Text("03")
+          }
+        }
+        item {
+          Button(
+            modifier = Modifier
+              .padding(24.dp)
+              .semantics { stateDescription = "State" },
+            onClick = {}
+          ) {
+            androidx.compose.material.Text("04")
+          }
+        }
+        item {
+          Button(
+            modifier = Modifier
+              .padding(24.dp)
+              .semantics { role = Role.DropdownList },
+            onClick = {}
+          ) {
+            androidx.compose.material.Text("05")
+          }
+        }
+        item {
+          Button(
+            modifier = Modifier
+              .padding(24.dp)
+              .clickable(onClickLabel = "Explore") {}
+              .semantics {
+                stateDescription = "State"
+                role = Role.RadioButton
+              },
+            onClick = {},
+            enabled = false
+          ) {
+            androidx.compose.material.Text("06")
+          }
+        }
+        item {
+          Button(
+            modifier = Modifier
+              .padding(24.dp)
+              .semantics {
+                selected = true
+                role = Role.Switch
+                heading()
+              },
+            onClick = {}
+          ) {
+            androidx.compose.material.Text("07")
+          }
+        }
+        item {
+          IconToggleButton(checked = true, onCheckedChange = { }) {
+            androidx.compose.material.Text("Toggle")
+          }
+        }
+      }
     }
+  }
 }
