+++ sample/build.gradle
@@ -30,6 +30,7 @@ androidComponents {
 dependencies {
   implementation libs.composeUi.material
   implementation libs.composeUi.material.icons
+  implementation libs.composeUi.material3.android
   implementation libs.composeUi.uiTooling
   // Remove if https://github.com/airbnb/lottie-android/pull/2507 lands
   implementation libs.androidx.appcompat
