+++ sample/src/test/snapshots/images/app.cash.paparazzi.sample_AccessibilityRenderingTest_legendDoesNotScale.png
