+++ sample/src/test/snapshots/images/app.cash.paparazzi.sample_AccessibilityRenderingTest_verify_hidden_ComposeView_content_is_not_in_legend.png
