+++ sample/src/test/snapshots/images/app.cash.paparazzi.sample_AccessibilityRenderingTest_verify_hidden_views_are_not_in_legend.png
