+++ sample/src/test/snapshots/images/app.cash.paparazzi.sample_AccessibilityRenderingTest_verify_view_live_region.png
